+++ blade-engine/src/com/bladecoder/engine/model/SpriteActor.java
@@ -83,10 +83,6 @@ public class SpriteActor extends BaseActor {
 				
 				setScale(s);
 			}
-
-			if (scene.getCameraFollowActor() == this)
-				scene.getCamera().updatePos(this);
-
 		}
 
 	}